[#939] [#940] [#941] [#942] Cleanup of stray processes, NFS issues workaround, DXR integration, adjusting RPC timeout and using random PG ports in YSQL Java tests
Summary:
This revision contains multiple improvements of our build and test framework.

== Cleanup of stray processes ==

Adding a script for cleaning up stray processes that fail to stop after the completion of the test.
The cleanup script starts with the given pid, wakes up often and adds any children of the current
pid set to that pid set, as well as removes any terminated pids. In the end it kills any processes
that have not terminated, not including the initial ancestor pid. `run-test.sh` starts this cleanup
script in the background, and before stopping, it sends the cleanup script a SIGUSR1 and processes
the cleanup script's termination and its logs. The cleanup script is also being careful not to kill itself,
because it is also a part of the process tree that it is monitoring. In a further revision we plan to
adjust JUnit-compatible XML files to treat stray processes as errors.

== Workarounds for NFS-related issues ==

`run_tests_on_spark.py` is being enhanced to introduce a timeout of 30 seconds for its
child process, `run-test.sh`, to do all the preparatory work and start running the underlying C++
or Java test. If that does not happen (which has been observed in our macOS builds in a small
percentage of cases), the `run-test.sh` child process is terminated and an exception is raised,
so that the Spark scheduler can re-run this test, potentially on a different node. The cause of these
cases is still being investigated, but it is most likely NFS-related.

Another NFS-related workaround is in the `compiler-wrapper.sh` script, which, among other
things, is responsible for picking a worker node to run the compiler on using ssh. This is our
current approach to parallelizing C++ compilation, possible alternatives including distcc and icecc.
The NFS file with the list of worker nodes might not exist sometimes, and, potentially due to
negative NFS caching (which we will try to turn off in our infrastructure separately), this file
appears as nonexistent at times. In those cases we are now using a locally cached version of
the file.

== DXR integration ==

Set environment variables required for DXR (https://github.com/mozilla/dxr) indexing based on YB_...
versions of these variables that get passed to a distributed pool of workers running tests.

Also rewrite compiler command line for PostgreSQL code to reference source files in the top-level
`src/postgres` directory rather than those copied to `build/../postgres_build`. This improves DXR
indexing of PostgreSQL code. This will also enable code comprehension on PostgreSQL code within the
`src/postgres` directory for tools based on `compile_commands.json` (e.g.  Visual Studio Code with
clangd).

== Stability of PostgreSQL tests ==

Improve the logic for setting `retryable_rpc_single_call_timeout_ms` in PostgreSQL Java tests based on
the build type. Specifically, the timeout needs to be greater than 10 seconds on macOS debug builds.

Use random ports for PostgreSQL in Java tests. This reduces the probability of port collisions on macOS,
where we use random loopback IP addresses of the form `127.0.0.x` (not `127.x.y.z` as we do on Linux).

== Other changes ==

- Add a flag `--use_libbacktrace` that allows using `libbacktrace` for symbolizing stack traces even in release mode, which is useful in tests.
- Remove the ability to build LLVM 6.x in thirdparty, because DXR now supports the Clang 7 plugin API.
- Treat non-existence of a third-party dependency's source directory as a reason to rebuild that dependency.

Test Plan:
- Jenkins: platform: linux
- Perform DXR indexing manually

Reviewers: sergei, timur, robert, bharat

Reviewed By: bharat

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D6230

+++ build-support/common-build-env.sh
@@ -82,6 +82,10 @@ if [[ -z ${is_run_test_script:-} ]]; then
 fi
 readonly is_run_test_script
 
+# Setting this to "true" will prevent any changes to the virtualenv (creating it or installing
+# modules into it) as part of activate_virtualenv.
+yb_readonly_virtualenv=false
+
 # -------------------------------------------------------------------------------------------------
 # Functions used in initializing some constants
 # -------------------------------------------------------------------------------------------------
@@ -1995,7 +1999,8 @@ activate_virtualenv() {
     fatal "Internal error: virtualenv_dir ('$virtualenv_dir') must end" \
           "with YB_VIRTUALENV_BASENAME ('$YB_VIRTUALENV_BASENAME')"
   fi
-  if [[ ${YB_RECREATE_VIRTUALENV:-} == "1" && -d $virtualenv_dir ]]; then
+  if [[ ${YB_RECREATE_VIRTUALENV:-} == "1" && -d $virtualenv_dir ]] && \
+     ! "$yb_readonly_virtualenv"; then
     log "YB_RECREATE_VIRTUALENV is set, deleting virtualenv at '$virtualenv_dir'"
     rm -rf "$virtualenv_dir"
     unset YB_RECREATE_VIRTUALENV
@@ -2005,6 +2010,9 @@ activate_virtualenv() {
   add_python_wrappers_dir_to_path
 
   if [[ ! -d $virtualenv_dir ]]; then
+    if "$yb_readonly_virtualenv"; then
+      fatal "virtualenv does not exist at '$virtualenv_dir', and we are not allowed to create it"
+    fi
     if [[ -n ${VIRTUAL_ENV:-} && -f $VIRTUAL_ENV/bin/activate ]]; then
       local old_virtual_env=$VIRTUAL_ENV
       # Re-activate and deactivate the other virtualenv we're in. Otherwise the deactivate
@@ -2026,6 +2034,7 @@ activate_virtualenv() {
       python2 -m virtualenv "$YB_VIRTUALENV_BASENAME"
     )
   fi
+
   set +u
   . "$virtualenv_dir"/bin/activate
   set -u
@@ -2034,8 +2043,10 @@ activate_virtualenv() {
     pip_no_cache="--no-cache-dir"
   fi
 
-  run_with_retries 10 0.5 pip2 install -r "$YB_SRC_ROOT/python_requirements_frozen.txt" \
-    $pip_no_cache
+  if ! "$yb_readonly_virtualenv"; then
+    run_with_retries 10 0.5 pip2 install -r "$YB_SRC_ROOT/python_requirements_frozen.txt" \
+      $pip_no_cache
+  fi
 }
 
 check_python_interpreter_version() {
@@ -2232,6 +2243,14 @@ set_java_home() {
 # -------------------------------------------------------------------------------------------------
 
 detect_os
+
+# http://man7.org/linux/man-pages/man7/signal.7.html
+if is_mac; then
+  declare -i -r SIGUSR1_EXIT_CODE=158  # 128 + 30
+else
+  # Linux
+  declare -i -r SIGUSR1_EXIT_CODE=138  # 128 + 10
+fi
 
 # This script is expected to be in build-support, a subdirectory of the repository root directory.
 set_yb_src_root "$( cd "$( dirname "$BASH_SOURCE" )"/.. && pwd )"
